INS-126: rearrange storage test files to fix coverage stats
